+++ database/migrations/2022_06_02_113223_create_withdraw_requests_table.php
@@ -1,0 +1,39 @@
+<?php
+
+use App\Models\User;
+use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class CreateWithdrawRequestsTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::create('withdraw_requests', function (Blueprint $table) {
+            $table->id();
+            $table->foreignIdFor(User::class);
+            $table->double('amount', 13,3);
+            $table->tinyInteger('status')->unsigned();
+            $table->string('message', 255)->nullable();
+            $table->bigInteger('approved_by')->nullable();
+            $table->dateTime('approved_at')->nullable();
+            $table->string('approval_note')->nullable();
+            $table->timestamps();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::dropIfExists('withdraw_requests');
+    }
+}