+++ resources/views/admin/layouts/sidebar.blade.php
@@ -59,6 +59,7 @@
                 <li><a href="{{ route('admin.category.index') }}"><i class="fa fa-list"></i>Job Category </a></li>
                 <li><a href="{{ route('user.verification.index') }}"><i class="fa fa-user"></i>Pending User </a>
                 <li><a href="{{ route('contract.index') }}"><i class="fa fa-dollar"></i>Contract Dispute </a>
+                <li><a href="{{ route('admin.withdraw.request.index') }}"><i class="fa fa-circle-dollar-to-slot"></i>Withdraw requests </a>
                 </li>
                 {{-- <li><a><i class="fa fa-bar-chart-o"></i> Data Presentation <span class="fa fa-chevron-down"></span></a>
                     <ul class="nav child_menu">