+++ resources/views/admin/user/list-user.blade.php
@@ -25,7 +25,6 @@
                 <div class="row">
                     <div class="col-sm-12">
                         <div class="card-box table-responsive">
-
                             <table id="datatable-buttons" class="table table-striped table-bordered" style="width:100%">
                                 <thead>
                                     <tr>
@@ -33,7 +32,6 @@
                                         <th>User Name</th>
                                         <th>Email</th>
                                         <th>Active Status</th>
-
                                         <th>Actions</th>
                                     </tr>
                                 </thead>